Also needed to move the models to core as well
