+++ src/NewRelic.Core/JsonConverters/BedrockPayloads/PayloadInterfaces.cs
@@ -1,7 +1,7 @@
 // Copyright 2020 New Relic, Inc. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-namespace NewRelic.Providers.Wrapper.Bedrock.Payloads
+namespace NewRelic.Core.JsonConverters.BedrockPayloads
 {
     public interface IRequestPayload
     {
